Backend/requirements.txt
asn1crypto==1.5.1
bip32==4.0
bip39==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coincurve==20.0.0
ecdsa==0.19.0
idna==3.10
mnemonic==0.21
pycparser==2.22
requests==2.32.3
six==1.17.0
urllib3==2.3.0
flask
flask-cors
python-dotenv
requests
